Move MachineLogsPublisher out of Kubernetes internal runtime
